+++ src/main/java/com/example/appointback/controller/TestObjectRepository.java
@@ -4,7 +4,7 @@ import com.example.appointback.entity.TestObject;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
-public interface TestObjectRepository extends CrudRepository<TestObject, Long> {
+public interface TestObjectRepository extends CrudRepository<TestObject, Integer> {
 
     List<TestObject> findAll();
 }
